pyproject.toml
[project]
name = "yurt_market"
version = "0.1.0"
description = "Multi-dorm student marketplace built with Django, DRF, and a modular clean architecture."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Django>=5.1,<6.0",
    "djangorestframework>=3.15.2,<4.0",
    "djangorestframework-simplejwt>=5.3.1,<6.0",
    "django-environ>=0.11.2",
    "django-cors-headers>=4.5.0,<5.0",
    "psycopg2-binary>=2.9.9,<3.0",
    "redis>=5.1.0",
    "celery>=5.4.0",
    "drf-spectacular>=0.27.2",
    "django-filter>=24.3",
    "phonenumberslite>=8.13.37",
    "Pillow>=10.4.0",
    "python-slugify>=8.0.4",
    "structlog>=24.1.0",
    "sentry-sdk>=2.16.0"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7.1",
    "pytest>=8.3.3",
    "pytest-django>=4.9.0",
    "factory-boy>=3.3.1",
    "model-bakery>=1.18.0",
    "django-debug-toolbar>=4.4.6"
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["config", "core", "modules", "scripts"]
exclude = ["tests", "tests.*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "DJ", "Q"]
ignore = ["E203", "E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

